Adding pytest-vcr for testing
